requirements.txt
asgiref==3.11.0
Django==5.0.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.7
pycountry==24.6.1
sqlparse==0.5.4
tzdata==2025.2
python-dotenv==1.2.1
django-phonenumber-field==8.4.0